Renamed MercureUpdatesGenerator to PublishingUpdatesGenerator to make it general purpose
