+++ module/Core/src/EventDispatcher/PublishingUpdatesGeneratorInterface.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\Mercure;
+namespace Shlinkio\Shlink\Core\EventDispatcher;
 
 use Shlinkio\Shlink\Common\UpdatePublishing\Update;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Entity\Visit;
 
-interface MercureUpdatesGeneratorInterface
+interface PublishingUpdatesGeneratorInterface
 {
     public function newVisitUpdate(Visit $visit): Update;
